requirements.txt
fastapi
uvicorn[standard]
pydantic
scikit-learn
numpy
pandas
python-multipart
python-jose[cryptography]
passlib[bcrypt]
